+++ crates/biome_fs/src/fs/os.rs
@@ -123,6 +123,26 @@ impl FileSystem for OsFileSystem {
             .map(|l| l.to_string())
             .collect())
     }
+
+    fn get_staged_files(&self) -> io::Result<Vec<String>> {
+        let output = Command::new("git")
+            .arg("diff")
+            .arg("--name-only")
+            .arg("--relative")
+            .arg("--staged")
+            // A: added
+            // C: copied
+            // M: modified
+            // R: renamed
+            // Source: https://git-scm.com/docs/git-diff#Documentation/git-diff.txt---diff-filterACDMRTUXB82308203
+            .arg("--diff-filter=ACMR")
+            .output()?;
+
+        Ok(String::from_utf8_lossy(&output.stdout)
+            .lines()
+            .map(|l| l.to_string())
+            .collect())
+    }
 }
 
 struct OsFile {